pyproject.toml
[project]
name = "tech-news-aggregator"
version = "0.1.0"
description = "Tech News Aggregation Agent System with LangGraph"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Tech News Team"}
]
dependencies = [
    # Web Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # LangGraph and LangChain
    "langgraph>=0.2.0",
    "langchain>=0.2.0",
    "langchain-openai>=0.1.0",

    # Database
    "sqlalchemy>=2.0.0",
    "sqlmodel>=0.0.14",
    "asyncpg>=0.29.0",
    "pgvector>=0.2.0",
    "alembic>=1.13.0",

    # RSS and Web Scraping
    "feedparser>=6.0.10",
    "trafilatura>=1.6.0",
    "httpx>=0.26.0",
    "beautifulsoup4>=4.12.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
    "structlog>=24.1.0",
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]